Put bar graph into seperate file, added time percentage graph start, will be a pie chart
