+++ src/graphs/TimePlayed.js
@@ -1,0 +1,11 @@
+
+const TimePlayed = ({games}) =>{
+
+
+    return(
+        <div>
+
+        </div>
+    );
+}
+export default TimePlayed;
